+++ packages/common/src/Config.ts
@@ -60,6 +60,8 @@ export interface CCFConfig {
     }
     RESOURCE_TAG_NAMES?: string[]
     CONSUMPTION_CHUNKS_DAYS?: number
+    SUBSCRIPTION_CHUNKS?: number
+    SUBSCRIPTIONS: string[]
   }
   ALI?: {
     NAME?: string
@@ -141,6 +143,12 @@ const getAzureResourceTagNames = () => {
 
 const getGCPProjects = () => {
   return process.env.GCP_PROJECTS ? process.env.GCP_PROJECTS : '[]'
+}
+
+const getAzureSubscriptions = () => {
+  return process.env.AZURE_SUBSCRIPTIONS
+    ? process.env.AZURE_SUBSCRIPTIONS
+    : '[]'
 }
 
 // This function allows support for using Docker Secrets.
@@ -275,6 +283,10 @@ const getConfig = (): CCFConfig => ({
     CONSUMPTION_CHUNKS_DAYS: parseInt(
       getEnvVar('AZURE_CONSUMPTION_CHUNKS_DAYS') || '0',
     ),
+    SUBSCRIPTION_CHUNKS: parseInt(
+      getEnvVar('AZURE_SUBSCRIPTION_CHUNKS') || '10',
+    ),
+    SUBSCRIPTIONS: JSON.parse(getAzureSubscriptions()) || [],
   },
   ALI: {
     NAME: 'ALI',