+++ packages/create-app/templates/default-app/packages/api/src/server.ts
@@ -11,18 +11,25 @@ import helmet from 'helmet'
 import cors, { CorsOptions } from 'cors'
 
 import { createRouter } from './api'
-import auth from './auth'
-import { Logger } from '@cloud-carbon-footprint/common'
+import { Logger, configLoader } from '@cloud-carbon-footprint/common'
+import { MongoDbCacheManager } from '@cloud-carbon-footprint/app'
+import swaggerDocs from './utils/swagger'
+import auth from './utils/auth'
 
 const port = process.env.PORT || 4000
 const httpApp = express()
-const serverLogger = new Logger('server')
+const serverLogger = new Logger('Server')
 
 if (process.env.NODE_ENV === 'production') {
   httpApp.use(auth)
 }
 
 httpApp.use(helmet())
+
+// Establish Mongo Connection if cache method selected
+if (configLoader()?.CACHE_MODE === 'MONGODB') {
+  MongoDbCacheManager.createDbConnection()
+}
 
 if (process.env.ENABLE_CORS) {
   const corsOptions: CorsOptions = {
@@ -41,8 +48,19 @@ if (process.env.ENABLE_CORS) {
 
 httpApp.use('/api', createRouter())
 
-httpApp.listen(port, () =>
+httpApp.listen(port, () => {
   serverLogger.info(
     `Cloud Carbon Footprint Server listening at http://localhost:${port}`,
-  ),
-)
+  )
+  swaggerDocs(httpApp, Number(port))
+})
+
+// Instructions for graceful shutdown
+process.on('SIGINT', async () => {
+  if (configLoader()?.CACHE_MODE === 'MONGODB') {
+    await MongoDbCacheManager.mongoClient.close()
+    serverLogger.info('\nMongoDB connection closed')
+  }
+  serverLogger.info('Cloud Carbon Footprint Server shutting down...')
+  process.exit()
+})